refactor: simplify store imports
